Split public and protected APIs
